AuthenticationService/requirements.txt
flask
flask-jwt-extended
flask-bcrypt
flask-sqlalchemy
python-dotenv
psycopg2-binary
redis
flask-caching
